backend/requirements.txt
fastapi
pydantic_settings
uvicorn
sqlalchemy
#alembic
asyncpg
psycopg2-binary
slowapi
python-dotenv
pyjwt
#passlib
pwdlib[argon2]
#bcrypt==4.0.1
python-multipart
email-validator
uvicorn[standard]==0.35.0
alembic
